+++ src/components/tag/tag.stories.js
@@ -4,7 +4,6 @@ import { boolean, select } from '@storybook/addon-knobs/react';
 import { Tag, TextBody, Tooltip } from '../../index';
 
 const sizes = ['small', 'medium', 'large'];
-const colors = ['neutral', 'mint', 'violet', 'ruby', 'gold', 'aqua'];
 
 const TooltippedTag = Tooltip(Tag);
 
@@ -13,21 +12,16 @@ export default {
 };
 
 export const basic = () => (
-  <Tag
-    color={select('Color', colors, 'neutral')}
-    size={select('Size', sizes, 'medium')}
-    inverse={boolean('Inverse', false)}
-    disabled={boolean('Disabled', false)}
-  >
+  <Tag size={select('Size', sizes, 'medium')} disabled={boolean('Disabled', false)}>
     I am a tag
   </Tag>
 );
 
 export const clickable = () => (
   <Tag
-    color={select('Color', colors, 'neutral')}
-    inverse={boolean('Inverse', false)}
-    onLabelClick={() => console.log('Tag label clicked')}
+    element="button"
+    onClick={() => console.log('Tag label clicked')}
+    selected={boolean('Selected', false)}
     size={select('Size', sizes, 'medium')}
     disabled={boolean('Disabled', false)}
   >
@@ -37,9 +31,8 @@ export const clickable = () => (
 
 export const closable = () => (
   <Tag
-    color={select('Color', colors, 'neutral')}
-    inverse={boolean('Inverse', false)}
     onRemoveClick={() => console.log('Tag removed')}
+    selected={boolean('Selected', false)}
     size={select('Size', sizes, 'medium')}
     disabled={boolean('Disabled', false)}
   >
@@ -49,10 +42,10 @@ export const closable = () => (
 
 export const clickableClosable = () => (
   <Tag
-    color={select('Color', colors, 'neutral')}
-    inverse={boolean('Inverse', false)}
-    onLabelClick={() => console.log('Tag label clicked')}
+    element="button"
+    onClick={() => console.log('Tag label clicked')}
     onRemoveClick={() => console.log('Tag removed')}
+    selected={boolean('Selected', false)}
     size={select('Size', sizes, 'medium')}
     disabled={boolean('Disabled', false)}
   >
@@ -66,10 +59,9 @@ clickableClosable.story = {
 
 export const withTooltip = () => (
   <TooltippedTag
-    color={select('Color', colors, 'neutral')}
-    inverse={boolean('Inverse', false)}
     tooltip={<TextBody>I am the tooltip</TextBody>}
     disabled={boolean('Disabled', false)}
+    selected={boolean('Selected', false)}
   >
     Tag with tooltip
   </TooltippedTag>